+++ src/Enums/CmsOptionName.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Vdhicts\Cyberfusion\ClusterApi\Enums;
+
+class CmsOptionName
+{
+    public const BLOG_PUBLIC = 'blog_public';
+
+    public const AVAILABLE = [
+        self::BLOG_PUBLIC,
+    ];
+}