+++ src/Models/CmsOption.php
@@ -1,0 +1,57 @@
+<?php
+
+namespace Vdhicts\Cyberfusion\ClusterApi\Models;
+
+use Vdhicts\Cyberfusion\ClusterApi\Support\Arr;
+use Vdhicts\Cyberfusion\ClusterApi\Contracts\Model;
+use Vdhicts\Cyberfusion\ClusterApi\Support\Validator;
+use Vdhicts\Cyberfusion\ClusterApi\Enums\CmsOptionName;
+
+class CmsOption extends ClusterModel implements Model
+{
+    private string $name;
+    private mixed $value;
+
+    public function getName(): string
+    {
+        return $this->name;
+    }
+
+    public function setName(string $name): CmsOption
+    {
+        Validator::value($name)
+            ->valueIn(CmsOptionName::AVAILABLE)
+            ->validate();
+
+        $this->name = $name;
+
+        return $this;
+    }
+
+    public function getValue(): mixed
+    {
+        return $this->value;
+    }
+
+    public function setValue(mixed $value): CmsOption
+    {
+        $this->value = $value;
+
+        return $this;
+    }
+
+    public function fromArray(array $data): CmsOption
+    {
+        return $this
+            ->setName(Arr::get($data, 'name'))
+            ->setValue(Arr::get($data, 'value'));
+    }
+
+    public function toArray(): array
+    {
+        return [
+            'name' => $this->getName(),
+            'value' => $this->getValue(),
+        ];
+    }
+}